Prettified Dao subfolder (#864)

+++ contractsSol5/Dao/mock/MockChainLinkSanityRate.sol
@@ -4,13 +4,13 @@ import "../ISanityRate.sol";
 
 
 contract MockChainLinkSanityRate is ISanityRate {
-    uint latestAnswerValue;
+    uint256 latestAnswerValue;
 
-    function setLatestKncToEthRate(uint _kncEthRate) external {
+    function setLatestKncToEthRate(uint256 _kncEthRate) external {
         latestAnswerValue = _kncEthRate;
     }
 
-    function latestAnswer() external view returns (uint) {
+    function latestAnswer() external view returns (uint256) {
         return latestAnswerValue;
     }
 }
