+++ contractsSol5/Dao/mock/MockMaliciousFeeHandlerReentrancy.sol
@@ -5,10 +5,9 @@ import "../KyberDAO.sol";
 
 /// @notice Mock Malicious Fee Handler tries to en-enter claim reward in DAO
 contract MockMaliciousFeeHandlerReentrancy {
-
     KyberDAO public dao;
 
-    uint public numberCalls;
+    uint256 public numberCalls;
 
     constructor() public {}
 
@@ -16,11 +15,15 @@ contract MockMaliciousFeeHandlerReentrancy {
         dao = _dao;
     }
 
-    function setNumberCalls(uint _numberCalls) public {
+    function setNumberCalls(uint256 _numberCalls) public {
         numberCalls = _numberCalls;
     }
 
-    function claimStakerReward(address payable staker, uint, uint epoch) public returns(bool) {
+    function claimStakerReward(
+        address payable staker,
+        uint256,
+        uint256 epoch
+    ) public returns (bool) {
         if (numberCalls > 0) {
             numberCalls--;
             dao.claimReward(staker, epoch);
